replace cat
I have suggested two changes, the first to change the catch for the species to a stop rather than defaulting to TOA.  It caught me out earlier when I was looking for TOP and mistyped to TOO and the message was lost in output. 

Second change is to turn the cat() at the end to a message() this way people can choose to suppress the output if using it in markdowns.